refactor: split asset functionality into own hook
