+++ src/containers/other/NFTView/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useNftData } from './useNftData';
 export { useViewBidsData } from './useViewBidsData';
+export { useAsset } from './useAsset';
